+++ packages/block-editor/src/components/mobile/picker/index.android.js
@@ -9,7 +9,7 @@ import { View } from 'react-native';
  */
 import { __ } from '@wordpress/i18n';
 import { Component } from '@wordpress/element';
-import { BottomSheet } from '@wordpress/editor';
+import { BottomSheet } from '@wordpress/block-editor';
 
 export default class Picker extends Component {
 	constructor() {