+++ packages/block-library/src/button/index.js
@@ -7,7 +7,6 @@ import { omit, pick } from 'lodash';
 /**
  * WordPress dependencies
  */
-import { G, Path, SVG } from '@wordpress/components';
 import { __, _x } from '@wordpress/i18n';
 import {
 	RichText,
@@ -18,6 +17,7 @@ import {
  * Internal dependencies
  */
 import edit from './edit';
+import icon from './icon';
 
 const blockAttributes = {
 	url: {
@@ -66,7 +66,7 @@ export const settings = {
 
 	description: __( 'Prompt visitors to take action with a button-style link.' ),
 
-	icon: <SVG viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><Path fill="none" d="M0 0h24v24H0V0z" /><G><Path d="M19 6H5c-1.1 0-2 .9-2 2v8c0 1.1.9 2 2 2h14c1.1 0 2-.9 2-2V8c0-1.1-.9-2-2-2zm0 10H5V8h14v8z" /></G></SVG>,
+	icon,
 
 	category: 'layout',