+++ packages/block-library/src/separator/edit.js
@@ -1,0 +1,3 @@
+export default function SeparatorEdit( { className } ) {
+	return <hr className={ className } />;
+}