+++ packages/block-library/src/separator/index.js
@@ -3,7 +3,12 @@
  */
 import { __ } from '@wordpress/i18n';
 import { createBlock } from '@wordpress/blocks';
-import { SVG, Path } from '@wordpress/components';
+
+/**
+ * Internal dependencies
+ */
+import edit from './edit';
+import icon from './icon';
 
 export const name = 'core/separator';
 
@@ -12,7 +17,7 @@ export const settings = {
 
 	description: __( 'Create a break between ideas or sections with a horizontal separator.' ),
 
-	icon: <SVG viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><Path fill="none" d="M0 0h24v24H0V0z" /><Path d="M19 13H5v-2h14v2z" /></SVG>,
+	icon,
 
 	category: 'layout',
 
@@ -41,9 +46,7 @@ export const settings = {
 		],
 	},
 
-	edit( { className } ) {
-		return <hr className={ className } />;
-	},
+	edit,
 
 	save() {
 		return <hr />;