+++ packages/block-library/src/text-columns/index.js
@@ -8,17 +8,21 @@ import { get, times } from 'lodash';
  */
 import { createBlock } from '@wordpress/blocks';
 import { __ } from '@wordpress/i18n';
-import { PanelBody, RangeControl } from '@wordpress/components';
-import { Fragment } from '@wordpress/element';
-import {
-	BlockControls,
-	BlockAlignmentToolbar,
-	InspectorControls,
-	RichText,
-} from '@wordpress/block-editor';
-import deprecated from '@wordpress/deprecated';
+import { RichText } from '@wordpress/block-editor';
 
-export const name = 'core/text-columns';
+/**
+ * Internal dependencies
+ */
+import edit from './edit';
+
+/**
+ * Internal dependencies
+ */
+import metadata from './block.json';
+
+const { name } = metadata;
+
+export { metadata, name };
 
 export const settings = {
 	// Disable insertion as this block is deprecated and ultimately replaced by the Columns block.
@@ -29,32 +33,6 @@ export const settings = {
 	title: __( 'Text Columns (deprecated)' ),
 
 	description: __( 'This block is deprecated. Please use the Columns block instead.' ),
-
-	icon: 'columns',
-
-	category: 'layout',
-
-	attributes: {
-		content: {
-			type: 'array',
-			source: 'query',
-			selector: 'p',
-			query: {
-				children: {
-					type: 'string',
-					source: 'html',
-				},
-			},
-			default: [ {}, {} ],
-		},
-		columns: {
-			type: 'number',
-			default: 2,
-		},
-		width: {
-			type: 'string',
-		},
-	},
 
 	transforms: {
 		to: [
@@ -89,60 +67,7 @@ export const settings = {
 		}
 	},
 
-	edit: ( ( { attributes, setAttributes, className } ) => {
-		const { width, content, columns } = attributes;
-
-		deprecated( 'The Text Columns block', {
-			alternative: 'the Columns block',
-			plugin: 'Gutenberg',
-		} );
-
-		return (
-			<Fragment>
-				<BlockControls>
-					<BlockAlignmentToolbar
-						value={ width }
-						onChange={ ( nextWidth ) => setAttributes( { width: nextWidth } ) }
-						controls={ [ 'center', 'wide', 'full' ] }
-					/>
-				</BlockControls>
-				<InspectorControls>
-					<PanelBody>
-						<RangeControl
-							label={ __( 'Columns' ) }
-							value={ columns }
-							onChange={ ( value ) => setAttributes( { columns: value } ) }
-							min={ 2 }
-							max={ 4 }
-							required
-						/>
-					</PanelBody>
-				</InspectorControls>
-				<div className={ `${ className } align${ width } columns-${ columns }` }>
-					{ times( columns, ( index ) => {
-						return (
-							<div className="wp-block-column" key={ `column-${ index }` }>
-								<RichText
-									tagName="p"
-									value={ get( content, [ index, 'children' ] ) }
-									onChange={ ( nextContent ) => {
-										setAttributes( {
-											content: [
-												...content.slice( 0, index ),
-												{ children: nextContent },
-												...content.slice( index + 1 ),
-											],
-										} );
-									} }
-									placeholder={ __( 'New Column' ) }
-								/>
-							</div>
-						);
-					} ) }
-				</div>
-			</Fragment>
-		);
-	} ),
+	edit,
 
 	save( { attributes } ) {
 		const { width, content, columns } = attributes;