+++ packages/data/src/test/registry.js
@@ -555,7 +555,8 @@ describe( 'createRegistry', () => {
 				},
 			} );
 
-			registry.dispatch( 'counter' ).increment(); // state = 1
+			const dispatchResult = registry.dispatch( 'counter' ).increment(); // state = 1
+			expect( dispatchResult ).toBe( undefined ); // Actions are implementation detail.
 			registry.dispatch( 'counter' ).increment( 4 ); // state = 5
 			expect( store.getState() ).toBe( 5 );
 		} );
@@ -602,6 +603,54 @@ describe( 'createRegistry', () => {
 			registry.use( plugin, { value: 10 } );
 
 			expect( registry.select() ).toBe( 10 );
+		} );
+	} );
+
+	describe( 'parent registry', () => {
+		it( 'should call parent registry selectors/actions if defined', () => {
+			const mySelector = jest.fn();
+			const myAction = jest.fn();
+			const getSelectors = () => ( { mySelector } );
+			const getActions = () => ( { myAction } );
+			const subscribe = () => {};
+			registry.registerGenericStore( 'store', { getSelectors, getActions, subscribe } );
+			const subRegistry = createRegistry( {}, registry );
+
+			subRegistry.select( 'store' ).mySelector();
+			subRegistry.dispatch( 'store' ).myAction();
+
+			expect( mySelector ).toHaveBeenCalled();
+			expect( myAction ).toHaveBeenCalled();
+		} );
+
+		it( 'should override existing store in parent registry', () => {
+			const mySelector = jest.fn();
+			const myAction = jest.fn();
+			const getSelectors = () => ( { mySelector } );
+			const getActions = () => ( { myAction } );
+			const subscribe = () => {};
+			registry.registerGenericStore( 'store', { getSelectors, getActions, subscribe } );
+
+			const subRegistry = createRegistry( {}, registry );
+			const mySelector2 = jest.fn();
+			const myAction2 = jest.fn();
+			const getSelectors2 = () => ( { mySelector: mySelector2 } );
+			const getActions2 = () => ( { myAction: myAction2 } );
+			const subscribe2 = () => {};
+			subRegistry.registerGenericStore( 'store', {
+				getSelectors: getSelectors2,
+				getActions: getActions2,
+				subscribe: subscribe2,
+			} );
+
+			subRegistry.select( 'store' ).mySelector();
+			subRegistry.dispatch( 'store' ).myAction();
+
+			expect( mySelector ).not.toHaveBeenCalled();
+			expect( myAction ).not.toHaveBeenCalled();
+
+			expect( mySelector2 ).toHaveBeenCalled();
+			expect( myAction2 ).toHaveBeenCalled();
 		} );
 	} );
 } );