+++ packages/components/CHANGELOG.md
@@ -1,8 +1,10 @@
-## 7.2.1 (Unreleased)
+## 7.3.0 (Unreleased)
 
 ### New Features
 
 - Added a new `render` property to `FormFileUpload` component. Allowing users of the component to custom the UI for their needs.
+- Added a new `BaseControl.VisualLabel` component.
+- Added a new `preview` prop to the `Placeholder` component which allows to display a preview, for example a media preview when the Placeholder is used in media editing contexts.
 
 ### Bug fixes